various improvements to fix the odd case where models were being passed to get_prep_value
changed IntegerIdentifierValue to subclass object. Probably not
important but it is working and I don't want to change it in this
commit.  But it does seem more appropriate...

Added get_db_prep_value which seems to be the main fix of this commit.

Adjust tests.  obj.pk seems to return the pk