override random to return IntegerIdentifierValue instances
This allows the removal of the hackish contribute_to_class save wrapper
calling to_python.

Explicitly setting the field to an int will result in the original
behavior that prompted the hack.  However, after a followup discussion
on django developers group it has become clear to me that this is the
correct approach.

https://groups.google.com/forum/#!topic/django-developers/3m71r7x7Lag